backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
google-genai
pandas
python-multipart
sentence-transformers
scikit-learn
slowapi
